pyproject.toml
[project]
name = "agenticlanggraph"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arxiv>=2.2.0",
    "ipykernel>=6.29.5",
    "langchain>=0.3.26",
    "langchain-community>=0.3.27",
    "langchain-groq>=0.3.5",
    "langchain-tavily>=0.2.6",
    "langgraph>=0.5.0",
    "langgraph-cli[inmem]>=0.3.3",
    "langsmith>=0.4.4",
    "matplotlib>=3.10.3",
    "python-dotenv>=1.1.1",
    "wikipedia>=1.4.0",
]
